Use viewWillDisappear instead of viewDidDisappear to invalidate changes
This will avoid race conditions due to other observers being set on
the next view's viewWillAppear. The reason we use viewWillAppear
instead of viewDidAppear is that we want to set up timers and
observers as early as possible. Also, setting the contacts name
changes the contact name mid-segue if we use viewDidAppear. And we
can't use viewDidLoad b/c that might not be called every time the view
appears.