Fix subtle bug with auto accepting
We were not removing the CallManager as an observer after going
through the connection and map screens. As such, the CallManager
continued to automatically accept calls whenever it knew that there
was a call to be received. The solution is to either remove the
CallManager as an observer or to create a wrapper function for the
receiveConnection method so that removing self as an observer would
also stop auto-accepting calls. I opted for the latter solution.