Understood why reference must be 'strong'
As Chris explained, if you allocate the object programmatically, you
need to have a strong reference as the instance variable won't matter
right after you allocate it if you have a weak reference. That is, right
after its creation, it will be unusable if it were to be a weak
reference since no one else holds a reference to it at that point.